Add badge customizer to the endpoint page
