+++ frontend/components/customizer/copied-content-indicator.js
