+++ frontend/components/customizer/path-builder.spec.js
